Proper adaptor clean up, preliminary auto-addressing support
The long-term plan for automatic address assignment is probably that we should be an IPAM
plugin as well, and on the backend spawn a DHCP server (if requested) that removes DHCP
assigned addresses from the pool.

As it is, you specify a subnet range you want to assign in, and spawn a DHCP in a range you
promise not to conflict in.