Add vde-ipam driver to docker-vde-driver.
This simple IPAM driver allows multiple overlapping networks to be created
by Docker to support VDE network testing.

It also allows binding VDE containers to the gateway IP of the network.

Special-case handling for gateway IPs to allow the creation of bridging
containers via docker-compose. The Gateway IP may be assigned when it is not
going to be set as the default route of the container (as is the case for any
layer 2 bridging container).