Allow specifying vde_switch, vde_plug binaries. Switch to using vdeplug4.
vdeplug4 supports a much more versatile backend. This set of patches adds
initial cut-over support for it. The plan going forward will probably be to
ship a statically compiled version of vdeplug4 with the plugin, since it's
otherwise totally backwards compatible.

It might be worth also incorporating vde_switch. In both cases, the model will
be that on startup we extract and run from tmp.